add mmacontext

+++ mma/mmasolver.cpp
@@ -30,7 +30,7 @@
 // PUBLIC
 ////////////////////////////////////////////////////////////////////////////////
 
-MMASolver::MMASolver(int nn, int mm, double ai, double ci, double di)
+MMASolver::MMASolver(int nn, int mm, float ai, float ci, float di)
 	: n(nn)
 	, m(mm)
 	, iter(0)
@@ -63,7 +63,7 @@ MMASolver::MMASolver(int nn, int mm, double ai, double ci, double di)
 	, xold2(n)
 { }
 
-void MMASolver::SetAsymptotes(double init, double decrease, double increase) {
+void MMASolver::SetAsymptotes(float init, float decrease, float increase) {
 
 	// asymptotes initialization and increase/decrease
 	asyminit = init;
@@ -71,8 +71,8 @@ void MMASolver::SetAsymptotes(double init, double decrease, double increase) {
 	asyminc = increase;
 }
 
-void MMASolver::Update(double *xval, const double *dfdx, const double *gx, const double *dgdx,
-	const double *xmin, const double *xmax)
+void MMASolver::Update(float *xval, const float *dfdx, const float *gx, const float *dgdx,
+	const float *xmin, const float *xmax)
 {
 	// Generate the subproblem
 	GenSub(xval, dfdx, gx, dgdx, xmin, xmax);
@@ -92,16 +92,16 @@ void MMASolver::Update(double *xval, const double *dfdx, const double *gx, const
 // PRIVATE
 ////////////////////////////////////////////////////////////////////////////////
 
-void MMASolver::SolveDIP(double *x) {
+void MMASolver::SolveDIP(float *x) {
 
 	for (int j = 0; j < m; j++) {
 		lam[j] = c[j] / 2.0;
 		mu[j] = 1.0;
 	}
 
-	const double tol = epsimin; // 1.0e-9*sqrt(m+n);
-	double epsi = 1.0;
-	double err = 1.0;
+	const float tol = epsimin; // 1.0e-9*sqrt(m+n);
+	float epsi = 1.0;
+	float err = 1.0;
 	int loop;
 
 	while (epsi > tol) {
@@ -146,33 +146,33 @@ void MMASolver::SolveDIP(double *x) {
 	}
 }
 
-void MMASolver::SolveDSA(double *x) {
+void MMASolver::SolveDSA(float *x) {
 
 	for (int j = 0; j < m; j++) {
 		lam[j] = 1.0;
 	}
 
-	const double tol = epsimin; // 1.0e-9*sqrt(m+n);
-	double err = 1.0;
+	const float tol = epsimin; // 1.0e-9*sqrt(m+n);
+	float err = 1.0;
 	int loop = 0;
 
 	while (err > tol && loop < 500) {
 		loop++;
 		XYZofLAMBDA(x);
 		DualGrad(x);
-		double theta = 1.0;
+		float theta = 1.0;
 		err = 0.0;
 		for (int j = 0; j < m; j++) {
-			lam[j] = std::max(0.0, lam[j] + theta * grad[j]);
+			lam[j] = std::max(0.f, lam[j] + theta * grad[j]);
 			err += grad[j] * grad[j];
 		}
 		err = std::sqrt(err);
 	}
 }
 
-double MMASolver::DualResidual(double *x, double epsi) {
+float MMASolver::DualResidual(float *x, float epsi) {
 
-	double *res = new double[2 * m];
+	float *res = new float[2 * m];
 
 	for (int j = 0; j < m; j++) {
 		res[j] = -b[j] - a[j] * z - y[j] + mu[j];
@@ -182,7 +182,7 @@ double MMASolver::DualResidual(double *x, double epsi) {
 		}
 	}
 
-	double nrI = 0.0;
+	float nrI = 0.0;
 	for (int i = 0; i < 2 * m; i++) {
 		if (nrI < std::abs(res[i])) {
 			nrI = std::abs(res[i]);
@@ -196,7 +196,7 @@ double MMASolver::DualResidual(double *x, double epsi) {
 
 void MMASolver::DualLineSearch() {
 
-	double theta = 1.005;
+	float theta = 1.005;
 	for (int i = 0; i < m; i++) {
 		if (theta < -1.01 * s[i] / lam[i]) {
 			theta = -1.01 * s[i] / lam[i];
@@ -213,23 +213,23 @@ void MMASolver::DualLineSearch() {
 	}
 }
 
-void MMASolver::DualHess(double *x) {
+void MMASolver::DualHess(float *x) {
 
-	double *df2 = new double[n];
-	double *PQ = new double[n * m];
+	float *df2 = new float[n];
+	float *PQ = new float[n * m];
 	#ifdef MMA_WITH_OPENMP
 	#pragma omp parallel for
 	#endif
 	for (int i = 0; i < n; i++) {
-		double pjlam = p0[i];
-		double qjlam = q0[i];
+		float pjlam = p0[i];
+		float qjlam = q0[i];
 		for (int j = 0; j < m; j++) {
 			pjlam += pij[i * m + j] * lam[j];
 			qjlam += qij[i * m + j] * lam[j];
 			PQ[i * m + j] = pij[i * m + j] / pow(upp[i] - x[i], 2.0) - qij[i * m + j] / pow(x[i] - low[i], 2.0);
 		}
 		df2[i] = -1.0 / (2.0 * pjlam / pow(upp[i] - x[i], 3.0) + 2.0 * qjlam / pow(x[i] - low[i], 3.0));
-		double xp = (sqrt(pjlam) * low[i] + sqrt(qjlam) * upp[i]) / (sqrt(pjlam) + sqrt(qjlam));
+		float xp = (sqrt(pjlam) * low[i] + sqrt(qjlam) * upp[i]) / (sqrt(pjlam) + sqrt(qjlam));
 		if (xp < alpha[i]) {
 			df2[i] = 0.0;
 		}
@@ -239,7 +239,7 @@ void MMASolver::DualHess(double *x) {
 	}
 
 	// Create the matrix/matrix/matrix product: PQ^T * diag(df2) * PQ
-	double *tmp = new double[n * m];
+	float *tmp = new float[n * m];
 	for (int j = 0; j < m; j++) {
 		#ifdef MMA_WITH_OPENMP
 		#pragma omp parallel for
@@ -259,7 +259,7 @@ void MMASolver::DualHess(double *x) {
 		}
 	}
 
-	double lamai = 0.0;
+	float lamai = 0.0;
 	for (int j = 0; j < m; j++) {
 		if (lam[j] < 0.0) {
 			lam[j] = 0.0;
@@ -280,11 +280,11 @@ void MMASolver::DualHess(double *x) {
 	}
 
 	// pos def check
-	double HessTrace = 0.0;
+	float HessTrace = 0.0;
 	for (int i = 0; i < m; i++) {
 		HessTrace += hess[i * m + i];
 	}
-	double HessCorr = 1e-4 * HessTrace / m;
+	float HessCorr = 1e-4 * HessTrace / m;
 
 	if (-1.0 * HessCorr < 1.0e-7) {
 		HessCorr = -1.0e-7;
@@ -299,7 +299,7 @@ void MMASolver::DualHess(double *x) {
 	delete[] tmp;
 }
 
-void MMASolver::DualGrad(double *x) {
+void MMASolver::DualGrad(float *x) {
 	for (int j = 0; j < m; j++) {
 		grad[j] = -b[j] - a[j] * z - y[j];
 		for (int i = 0; i < n; i++) {
@@ -308,14 +308,14 @@ void MMASolver::DualGrad(double *x) {
 	}
 }
 
-void MMASolver::XYZofLAMBDA(double *x) {
+void MMASolver::XYZofLAMBDA(float *x) {
 
-	double lamai = 0.0;
+	float lamai = 0.0;
 	for (int i = 0; i < m; i++) {
 		if (lam[i] < 0.0) {
 			lam[i] = 0;
 		}
-		y[i] = std::max(0.0, lam[i] - c[i]); // Note y=(lam-c)/d - however d is fixed at one !!
+		y[i] = std::max(0.f, lam[i] - c[i]); // Note y=(lam-c)/d - however d is fixed at one !!
 		lamai += lam[i] * a[i];
 	}
 	z = std::max(0.0, 10.0 * (lamai - 1.0)); // SINCE a0 = 1.0
@@ -324,8 +324,8 @@ void MMASolver::XYZofLAMBDA(double *x) {
 	#pragma omp parallel for
 	#endif
 	for (int i = 0; i < n; i++) {
-		double pjlam = p0[i];
-		double qjlam = q0[i];
+		float pjlam = p0[i];
+		float qjlam = q0[i];
 		for (int j = 0; j < m; j++) {
 			pjlam += pij[i * m + j] * lam[j];
 			qjlam += qij[i * m + j] * lam[j];
@@ -340,8 +340,8 @@ void MMASolver::XYZofLAMBDA(double *x) {
 	}
 }
 
-void MMASolver::GenSub(const double *xval, const double *dfdx, const double *gx, const double *dgdx, const double *xmin,
-                       const double *xmax)
+void MMASolver::GenSub(const float *xval, const float *dfdx, const float *gx, const float *dgdx, const float *xmin,
+                       const float *xmax)
 {
 	// Forward the iterator
 	iter++;
@@ -360,8 +360,8 @@ void MMASolver::GenSub(const double *xval, const double *dfdx, const double *gx,
 		#pragma omp parallel for
 		#endif
 		for (int i = 0; i < n; i++) {
-			double zzz = (xval[i] - xold1[i]) * (xold1[i] - xold2[i]);
-			double gamma;
+			float zzz = (xval[i] - xold1[i]) * (xold1[i] - xold2[i]);
+			float gamma;
 			if (zzz < 0.0) {
 				gamma = asymdec;
 			} else if (zzz > 0.0) {
@@ -372,15 +372,15 @@ void MMASolver::GenSub(const double *xval, const double *dfdx, const double *gx,
 			low[i] = xval[i] - gamma * (xold1[i] - low[i]);
 			upp[i] = xval[i] + gamma * (upp[i] - xold1[i]);
 
-			double xmami = std::max(xmamieps, xmax[i] - xmin[i]);
-			// double xmami = xmax[i] - xmin[i];
-			low[i] = std::max(low[i], xval[i] - 100.0 * xmami);
-			low[i] = std::min(low[i], xval[i] - 1.0e-5 * xmami);
-			upp[i] = std::max(upp[i], xval[i] + 1.0e-5 * xmami);
-			upp[i] = std::min(upp[i], xval[i] + 100.0 * xmami);
+			float xmami = std::max(xmamieps, xmax[i] - xmin[i]);
+			// float xmami = xmax[i] - xmin[i];
+			low[i] = std::max(low[i], xval[i] - 100.f * xmami);
+			low[i] = std::min(low[i], xval[i] - 1.0e-5f * xmami);
+			upp[i] = std::max(upp[i], xval[i] + 1.0e-5f * xmami);
+			upp[i] = std::min(upp[i], xval[i] + 100.f * xmami);
 
-			double xmi = xmin[i] - 1.0e-6;
-			double xma = xmax[i] + 1.0e-6;
+			float xmi = xmin[i] - 1.0e-6;
+			float xma = xmax[i] + 1.0e-6;
 			if (xval[i] < xmi) {
 				low[i] = xval[i] - (xma - xval[i]) / 0.9;
 				upp[i] = xval[i] + (xma - xval[i]) / 0.9;
@@ -393,7 +393,7 @@ void MMASolver::GenSub(const double *xval, const double *dfdx, const double *gx,
 	}
 
 	// Set bounds and the coefficients for the approximation
-	// double raa0 = 0.5*1e-6;
+	// float raa0 = 0.5*1e-6;
 	#ifdef MMA_WITH_OPENMP
 	#pragma omp parallel for
 	#endif
@@ -408,20 +408,20 @@ void MMASolver::GenSub(const double *xval, const double *dfdx, const double *gx,
 
 		// Objective function
 		{
-			double dfdxp = std::max(0.0, dfdx[i]);
-			double dfdxm = std::max(0.0, -1.0 * dfdx[i]);
-			double xmamiinv = 1.0 / std::max(xmamieps, xmax[i] - xmin[i]);
-			double pq = 0.001 * std::abs(dfdx[i]) + raa0 * xmamiinv;
+			float dfdxp = std::max(0.f, dfdx[i]);
+			float dfdxm = std::max(0.f, -1.f * dfdx[i]);
+			float xmamiinv = 1.0 / std::max(xmamieps, xmax[i] - xmin[i]);
+			float pq = 0.001 * std::abs(dfdx[i]) + raa0 * xmamiinv;
 			p0[i] = std::pow(upp[i] - xval[i], 2.0) * (dfdxp + pq);
 			q0[i] = std::pow(xval[i] - low[i], 2.0) * (dfdxm + pq);
 		}
 
 		// Constraints
 		for (int j = 0; j < m; j++) {
-			double dgdxp = std::max(0.0, dgdx[i * m + j]);
-			double dgdxm = std::max(0.0, -1.0 * dgdx[i * m + j]);
-			double xmamiinv = 1.0 / std::max(xmamieps, xmax[i] - xmin[i]);
-			double pq = 0.001 * std::abs(dgdx[i * m + j]) + raa0 * xmamiinv;
+			float dgdxp = std::max(0.f, dgdx[i * m + j]);
+			float dgdxm = std::max(0.f, -1.f * dgdx[i * m + j]);
+			float xmamiinv = 1.0 / std::max(xmamieps, xmax[i] - xmin[i]);
+			float pq = 0.001 * std::abs(dgdx[i * m + j]) + raa0 * xmamiinv;
 			pij[i * m + j] = std::pow(upp[i] - xval[i], 2.0) * (dgdxp + pq);
 			qij[i * m + j] = std::pow(xval[i] - low[i], 2.0) * (dgdxm + pq);
 		}
@@ -436,7 +436,7 @@ void MMASolver::GenSub(const double *xval, const double *dfdx, const double *gx,
 	}
 }
 
-void MMASolver::Factorize(double *K, int n) {
+void MMASolver::Factorize(float *K, int n) {
 
 	for (int s = 0; s < n - 1; s++) {
 		for (int i = s + 1; i < n; i++) {
@@ -448,10 +448,10 @@ void MMASolver::Factorize(double *K, int n) {
 	}
 }
 
-void MMASolver::Solve(double *K, double *x, int n) {
+void MMASolver::Solve(float *K, float *x, int n) {
 
 	for (int i = 1; i < n; i++) {
-		double a = 0.0;
+		float a = 0.0;
 		for (int j = 0; j < i; j++) {
 			a = a - K[i * n + j] * x[j];
 		}
@@ -460,7 +460,7 @@ void MMASolver::Solve(double *K, double *x, int n) {
 
 	x[n - 1] = x[n - 1] / K[(n - 1) * n + (n - 1)];
 	for (int i = n - 2; i >= 0; i--) {
-		double a = x[i];
+		float a = x[i];
 		for (int j = i + 1; j < n; j++) {
 			a = a - K[i * n + j] * x[j];
 		}
